+++ src/loaders/utils.js
@@ -1,3 +1,5 @@
+import quickselect from 'quickselect';
+
 export function isInTileBounds({
   x,
   y,
@@ -59,4 +61,61 @@ export function byteSwapInplace(src) {
       flipper[i + endByteIndex - j] = t;
     }
   }
+}
+
+/**
+ * Returns actual image stats for static imagery and an estimate via a downsampled version of image pyramids.
+ * This is helpful for generating histograms of your channel data, or scaling your sliders down to a reasonable range.
+ * @param {Object} args
+ * @param {Object} args.loader A valid loader object.
+ * @param {Array} args.loaderSelection Array of valid dimension selections
+ * @returns {Array} List of { mean, domain, sd, data, q1, q3 } objects.
+ */
+export async function getChannelStats({ loader, loaderSelection }) {
+  const z = loader.isPyramid ? loader.numLevels - 1 : 0;
+  const rasters = await loader.getRaster({ z, loaderSelection });
+  const { data } = rasters;
+  const channelStats = data.map(arr => {
+    let len = arr.length;
+    let min = Infinity;
+    let max = -Infinity;
+    let total = 0;
+    // Range (min/max).
+    // eslint-disable-next-line no-plusplus
+    while (len--) {
+      if (arr[len] < min) {
+        min = arr[len];
+      }
+      if (arr[len] > max) {
+        max = arr[len];
+      }
+      total += arr[len];
+    }
+
+    // Mean.
+    const mean = total / arr.length;
+
+    // Standard Deviation.
+    len = arr.length;
+    let sumSquared = 0;
+    // eslint-disable-next-line no-plusplus
+    while (len--) {
+      sumSquared += (arr[len] - mean) ** 2;
+    }
+    const sd = (sumSquared / arr.length) ** 0.5;
+
+    // Median, and quartiles via quickselect: https://en.wikipedia.org/wiki/Quickselect.
+    // Odd number lengths should round down the index.
+    const mid = Math.floor(arr.length / 2);
+    const firstQuartileLocation = Math.floor(arr.length / 4);
+    const thirdQuartileLocation = 3 * Math.floor(arr.length / 4);
+    quickselect(arr, mid);
+    const median = arr[mid];
+    quickselect(arr, firstQuartileLocation, 0, mid);
+    const q1 = arr[firstQuartileLocation];
+    quickselect(arr, thirdQuartileLocation, mid, arr.length - 1);
+    const q3 = arr[thirdQuartileLocation];
+    return { mean, sd, q1, q3, median, data: arr, domain: [min, max] };
+  });
+  return channelStats;
 }